requirements.txt
streamlit
pandas
scikit-learn
scikit-image
pillow
numpy
opencv-python-headless